+++ static/app/components/omniSearch/palette.tsx
@@ -1,31 +1,18 @@
-import {Fragment, useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import {Fragment, useEffect, useMemo, useRef, useState} from 'react';
 import {useNavigate} from 'react-router-dom';
 import styled from '@emotion/styled';
 import * as CommandPrimitive from 'cmdk';
 
 import {closeModal} from 'sentry/actionCreators/modal';
-import {useOmniActions} from 'sentry/components/omniSearch/useOmniActions';
-import {useOmniSearchState} from 'sentry/components/omniSearch/useOmniSearchState';
-import {
-  createDocIntegrationResults,
-  createIntegrationResults,
-  createMemberResults,
-  createPluginResults,
-  createProjectResults,
-  createSentryAppResults,
-  createTeamResults,
-  queryResults,
-} from 'sentry/components/search/sources/apiSource';
-import type {ResultItem} from 'sentry/components/search/sources/types';
 import {strGetFn} from 'sentry/components/search/sources/utils';
-import {IconDocs} from 'sentry/icons';
 import {IconDefaultsProvider} from 'sentry/icons/useIconDefaults';
 import {createFuzzySearch} from 'sentry/utils/fuzzySearch';
-import useApi from 'sentry/utils/useApi';
 import {useDebouncedValue} from 'sentry/utils/useDebouncedValue';
-import useOrganization from 'sentry/utils/useOrganization';
 
 import type {OmniAction} from './types';
+import {useApiDynamicActions} from './useApiDynamicActions';
+import {useOmniActions} from './useOmniActions';
+import {useOmniSearchState} from './useOmniSearchState';
 
 /**
  * Very basic palette UI using cmdk that lists all registered actions.
@@ -44,63 +31,9 @@ export function OmniSearchPalette() {
   const navigate = useNavigate();
   const inputRef = useRef<HTMLInputElement>(null);
 
-  const api = useApi();
-  const organization = useOrganization({allowNull: true});
-
-  const [results, setResults] = useState<ResultItem[]>([]);
-
   const debouncedQuery = useDebouncedValue(query, 300);
 
-  const handleSearch = useCallback(async () => {
-    const pendingResults: Array<Promise<ResultItem[] | null>> = [];
-
-    if (organization) {
-      const org = organization;
-      const slug = organization.slug;
-
-      const q = (url: string) => queryResults(api, url, debouncedQuery);
-
-      const searchQueries = [
-        createProjectResults(q(`/organizations/${slug}/projects/`), org),
-        createTeamResults(q(`/organizations/${slug}/teams/`), org),
-        createMemberResults(q(`/organizations/${slug}/members/`), org),
-        createPluginResults(q(`/organizations/${slug}/plugins/configs/`), org),
-        createIntegrationResults(q(`/organizations/${slug}/config/integrations/`), org),
-        createSentryAppResults(q('/sentry-apps/?status=published'), org),
-        createDocIntegrationResults(q('/doc-integrations/'), org),
-      ];
-      pendingResults.push(...searchQueries);
-    }
-
-    const resolvedResults = await Promise.all(pendingResults);
-    setResults(resolvedResults.flat().filter(i => i !== null));
-  }, [api, debouncedQuery, organization]);
-
-  useEffect(() => {
-    void handleSearch();
-  }, [handleSearch]);
-
-  const dynamicActions = useMemo(() => {
-    const actions: OmniAction[] = [];
-    if (query) {
-      results.forEach((result, index) => {
-        actions.push({
-          key: `api-${index}`,
-          areaKey: 'navigate',
-          label: result.title as string,
-          actionIcon: <IconDocs />,
-          onAction: () => {
-            if (typeof result.to === 'string') {
-              window.open(result.to, '_blank', 'noreferrer');
-            }
-          },
-        });
-      });
-    }
-
-    return actions.slice(0, 10);
-  }, [results, query]);
-
+  const dynamicActions = useApiDynamicActions(debouncedQuery);
   useOmniActions(dynamicActions);
 
   const [filteredAvailableActions, setFilteredAvailableActions] = useState<OmniAction[]>(
@@ -117,7 +50,7 @@ export function OmniSearchPalette() {
   }, [availableActions, debouncedQuery]);
 
   const grouped = useMemo(() => {
-    // const actions = availableActions.filter((a: OmniAction) => !a.hidden);
+    // Filter actions based on query
     const actions = debouncedQuery ? filteredAvailableActions : availableActions;
 
     // Group by section label
